+++ EstimoteSDK/EstimoteSDK.framework/Versions/A/Headers/ESTRequestPatchJSON.h
@@ -25,6 +25,6 @@ typedef NS_ENUM(NSInteger, ESTRequestPatchJSONError)
 
 @interface ESTRequestPatchJSON : ESTRequestBase
 
-- (void)setParams:(NSDictionary *)params forRequest:(NSMutableURLRequest *)request;
+- (void)setParams:(id)params forRequest:(NSMutableURLRequest *)request;
 
 @end