+++ EstimoteSDK/EstimoteSDK.framework/Versions/A/Headers/EstimoteSDK.h
@@ -7,7 +7,7 @@
 //  |______|___/\__|_|_| |_| |_|\___/ \__\___| |_____/|_____/|_|\_\
 //
 //
-//  Version: 3.5.0
+//  Version: 3.6.0
 //  Copyright (c) 2015 Estimote. All rights reserved.
 
 /**
@@ -71,12 +71,15 @@
 
 #import "ESTRequestGetBeacons.h"
 #import "ESTRequestBeaconDetails.h"
+#import "ESTRequestGetBeaconsDetails.h"
 #import "ESTRequestBeaconColor.h"
 #import "ESTRequestBeaconMac.h"
 #import "ESTRequestAssignGPSLocation.h"
 
 #import "ESTRequestRegisterDevice.h"
 #import "ESTRequestGetPendingSettings.h"
+#import "ESTRequestCancelPendingSettings.h"
+#import "ESTRequestGetSettingsHistory.h"
 
 #import "ESTRequestGetNearables.h"